Modif: dtos cambio type var

+++ dto/dto_product.go
@@ -4,7 +4,7 @@ type ProductDto struct {
 	IdProduct    string  `json:"id_product"`
 	Name_product string  `json:"name_product"`
 	Cost         float64 `json:"cost"`
-	Stock        float64 `json:"stock"`
+	Stock        int     `json:"stock"`
 	Category     string  `json:"category"`
 
 	/*Name string `json:"name"`
